@@ -6,12 +6,12 @@ class ExpandedGridContent extends StatelessWidget {
   final Widget child;
 
   ExpandedGridContent({
-    Key key,
+    Key? key,
     this.columnIndex = 0,
     this.rowIndex = 0,
     this.columnSpan = 1,
     this.rowSpan = 1,
-    @required this.child
+    required this.child
   }):super(key: key);
 
   @override
